Split uri manipulation
