+++ uri/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./main";
+export * from "./uri-manipulation";
